changed the BigDecimal objects to Integer

+++ TrackForce/src/main/java/com/revature/utils/PersistentStorage.java
@@ -71,7 +71,7 @@ public class PersistentStorage {
 	}
 
 	//added method to actually get values from the cache
-	public ClientInfo getClient(BigDecimal id) {
+	public ClientInfo getClient(Integer id) {
 		return clients.get(id);
 	}
